Allow UI to restore to current state
There is a momentarily flash of a Steam spinner while the state loads
from the backend -- it would be nice to fix this in future.

Additionally, after initialising state, the frontend reapplies the
current configuration. It would be good to stop sending requests when
the state hasn't changed, which probably needs browser storage...
(Although keeping the UI dumb lets the backend drive which has its own
advantages.)